LayoutTests/http/tests/quicklook/cross-origin-iframe-blocked.html

<!DOCTYPE html>
<html>
<head>
<script src="resources/tap-at-point-and-notify-done.js"></script>
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.dumpChildFramesAsText();
    testRunner.waitUntilDone();
}
</script>
</head>
<body>
<p>This test verifies that loading a cross-origin iframe is blocked when created by a QuickLook preview of a Microsoft Word document. This test PASSED only if a security error is logged to the console.</p>
<iframe src="resources/cross-origin-iframe-blocked.docx" onload="tapAtPointAndNotifyDone(this.offsetLeft + 5, this.offsetTop + 5)"></iframe>
</body>
</html>
